requirements.txt
django
djangoRestFramework
django-cors-headers
pyyaml
requests